requirements.txt
pinecone-client
langchain
pypdf
sentence_transformers
replicate